requirements.txt
# Python Version 3.10.12
fastapi==0.100.1
pymongo==4.6.0
pyngrok==7.0.1
requests==2.31.0
uvicorn==0.21.1
uvloop==0.19.0
python-dotenv==1.0.0